requirements.txt
Flask>=3.0,<4.0
PyMySQL>=1.1,<2.0
pandas>=2.2,<3.0
yfinance>=0.2.50,<0.3
python-dotenv>=1.0,<2.0
